Добавить удаление users_registration #249

+++ services/RegistrationDispatcher.php
@@ -12,8 +12,10 @@ class RegistrationDispatcher extends \cs\services\dispatcher\Registration
     public static function cron($isEcho = true)
     {
         $ids = (new Query())->select('parent_id')->from(static::TABLE)->where(['<', 'date_finish', gmdate('YmdHis')])->column();
-        \Yii::info(\yii\helpers\VarDumper::dumpAsString($ids), 'gs\\app\\services\\RegistrationDispatcher::cron');
-        \app\models\User::deleteByCondition(['in', 'id', $ids]);
+        if (count($ids) > 0) {
+            \Yii::info(\yii\helpers\VarDumper::dumpAsString($ids), 'gs\\app\\services\\RegistrationDispatcher::cron');
+            \app\models\User::deleteByCondition(['in', 'id', $ids]);
+        }
 
         parent::cron($isEcho);
     }
